+++ waterfurnace2/__init__.py
@@ -1,0 +1,7 @@
+# -*- coding: utf-8 -*-
+
+"""Top-level package for waterfurnace."""
+
+__author__ = """Sean Dague"""
+__email__ = 'sean@dague.net'
+__version__ = '1.1.0'